FEAT: Add user interface with menu
This commit adds a new root-node such that the addition of the user
interface becomes simpler. This has also resulted in the world node
now being a child of this new root node and can now also be replaced
for testing new worlds.

The added UI has a small menu added which currently moves the previous
quit option from "ESQ" to the "Quit"-button. "ESQ" now toggle the menu.

Restarting via "R" is currentlo not available
The "Start"-button does not reset the position of the player correctly