FEAT: Add option to restart the level
This commit adds the possibility to restart the level either via the
menu or pressing "R" on the keyboard.
Furthermore this commit also now pauses the game when the menu is
brought up using the "ESQ" key.

Gravity has also been re-enabled albeit with greatly reduced effect
on the character. Currently the gravity scale is set to 0.125.