FEAT: Add a stopwatch clock to the UI
This commit adds a stopwatch clock to the game and represents this in
the UI to the top left. This stopwatch is used to show the time passed
since the current level started.

This commit also adds a display count of how many tries that has been
made on the current level.

In addition to the above the commit also adds a countdown timer which
is currently only mechanically used, ie it holds the game paused for
three seconds. There should be a UI element to display the remaining
time, however this has not been implemented yet.

This commit also contains minor code cleanup and addition of
documentation to some parts of the code