FEAT: Add zones and UI elements to support splits
This commit adds the ability to register time splits based on the
character passing a specific zone. This is added at an early stages and
might be changed later. Currently these zones are added together with
the level and are removed when the player enters them to avoid adding
multiple splits.
Together with this, there is also updates to the UI such that a split
time is added in the panel to the right. This UI panel is currently
being removed/re-added for every _process() called from game.gd. If
this is actually needed, well that needs to be determined.

Currently the time splits appear to share a reference to the general
date_time class, such that they do not save a specific time...