FIX: Store a copy of the current time
This commit fixes the previous issue that the time splits referred to
the current time, meaning they would reflect the clock that was running
and not infact store the time for when a time_split zone was entered.
This is now fixed by duplicating the resource, when calling get_time()
on the stop_watch.