FEAT: Extend the current level and add traps
This commit extends the current level with additional obstacles that
the player needs to navigate. It also adds two traps or moving
obstacles, the pendulum and the trap door, that will be encountered
as the player gets further into the level.

The pendulum is just a simple pendulum, currently implemented as an
animated body that moves back and forth. This object will push the
player away if it gets to close.

The trap door is a triggered mechanism that "falls" down when the
player enters the trigger zone. After it has triggered it will take
some time to re-arm, before it can be triggered again.

Currently the trap door requires the player to leave and enter the
trigger zone, before the trap can be triggered again, even if it has
been re-armed